+++ cmd/shortener/main.go
@@ -1,83 +1,20 @@
 package main
 
 import (
-	"io"
 	"log"
-	"math/rand"
 	"net/http"
-	"strings"
-	"time"
+
+	"http/internal/app" // Импортируем обработчики
 )
-
-var urlStorage = make(map[string]string)
-
-const letters = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
-
-
-func generateShortID() string {
-	rand.Seed(time.Now().UnixNano())
-	id := make([]byte, 6)
-	for i := range id {
-		id[i] = letters[rand.Intn(len(letters))]
-	}
-	return string(id)
-}
-
-
-func handleShortenURL(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		http.Error(w, "Only POST method is allowed", http.StatusBadRequest)
-		return 
-	}
-
-	body, err := io.ReadAll(r.Body) 
-	defer r.Body.Close()            
-
-	if err != nil {
-		http.Error(w, "Failed to read request body", http.StatusInternalServerError)
-		return
-	}
-
-	url := strings.TrimSpace(string(body))
-	if url == "" {
-		http.Error(w, "Empty URL", http.StatusBadRequest)
-		return
-	}
-
-	shortID := generateShortID()
-	urlStorage[shortID] = url
-
-	shortURL := "http://localhost:8080/" + shortID
-	w.Header().Set("Content-Type", "text/plain")
-	w.WriteHeader(http.StatusCreated)
-	w.Write([]byte(shortURL))
-}
-
-func handleRedirect(w http.ResponseWriter, r *http.Request) {
-	id := strings.TrimPrefix(r.URL.Path, "/")
-	if id == "" {
-		http.Error(w, "Invalid ID", http.StatusBadRequest)
-		return
-	}
-
-	originalURL, found := urlStorage[id]
-	if !found {
-		http.Error(w, "URL not found", http.StatusBadRequest)
-		return
-	}
-
-	w.Header().Set("Content-Type", "text/plain")
-	http.Redirect(w, r, originalURL, http.StatusTemporaryRedirect)
-}
 
 func main() {
 	mux := http.NewServeMux()
 
 	mux.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
 		if r.URL.Path == "/" && r.Method == http.MethodPost {
-			handleShortenURL(w, r)
+			app.HandleShortenURL(w, r)
 		} else if r.Method == http.MethodGet {
-			handleRedirect(w, r)
+			app.HandleRedirect(w, r)
 		} else {
 			http.Error(w, "Invalid request", http.StatusBadRequest)
 		}
@@ -89,4 +26,3 @@ func main() {
 		log.Fatal(err)
 	}
 }
-
